+++ R/dplyr_spark_table.R
@@ -16,7 +16,6 @@ collect.tbl_spark <- function(x, ...) {
 
 #' @export
 #' @importFrom dplyr sample_n
-#' @importFrom dbplyr add_op_single
 sample_n.tbl_spark <- function(tbl,
                                size,
                                replace = FALSE,
@@ -34,13 +33,18 @@ sample_n.tbl_spark <- function(tbl,
     .env = .env
   )
 
-  add_op_single("sample_n", .data = tbl, args = args) %>%
+  if (dbplyr_uses_ops()) {
+    tbl <- dbplyr::add_op_single("sample_n", .data = tbl, args = args)
+  } else {
+    tbl$lazy_query <- lazy_sample_query(tbl$lazy_query, frac = FALSE, args = args)
+  }
+
+  tbl %>%
     as_sampled_tbl(frac = FALSE, args = args)
 }
 
 #' @export
 #' @importFrom dplyr sample_frac
-#' @importFrom dbplyr add_op_single
 sample_frac.tbl_spark <- function(tbl,
                                   size = 1,
                                   replace = FALSE,
@@ -57,7 +61,14 @@ sample_frac.tbl_spark <- function(tbl,
     seed = gen_prng_seed(),
     .env = .env
   )
-  add_op_single("sample_frac", .data = tbl, args = args) %>%
+
+  if (dbplyr_uses_ops()) {
+    tbl <- dbplyr::add_op_single("sample_frac", .data = tbl, args = args)
+  } else {
+    tbl$lazy_query <- lazy_sample_query(tbl$lazy_query, frac = TRUE, args = args)
+  }
+
+  tbl %>%
     as_sampled_tbl(frac = TRUE, args = args)
 }
 
@@ -119,8 +130,16 @@ print.tbl_spark <- function(x, ...) {
     rows <- max(rows, options$n)
   }
 
-  grps <- dbplyr::op_grps(x$ops)
-  sort <- dbplyr::op_sort(x$ops) %>%
+  if (dbplyr_uses_ops()) {
+    grps <- dbplyr::op_grps(x$ops)
+    sort <- dbplyr::op_sort(x$ops)
+  } else {
+    # TODO should this use `$lazy_query`?
+    grps <- dbplyr::op_grps(x)
+    sort <- dbplyr::op_sort(x)
+  }
+
+  sort <- sort %>%
     purrr::map_if(rlang::is_formula, rlang::f_rhs) %>%
     purrr::map_chr(rlang::expr_text, width = 500L)
 
@@ -160,5 +179,5 @@ print.tbl_spark <- function(x, ...) {
 #' @export
 #' @importFrom dplyr tbl_ptype
 tbl_ptype.tbl_spark <- function(.data) {
-  simulate_vars(.data)
+  simulate_vars_spark(.data)
 }