+++ R/zzz.R
@@ -7,4 +7,11 @@ set_option_default <- function(...) {
   })
 }
 
+.onLoad <- function(...) {
+  if (!dbplyr_uses_ops()) {
+    vctrs::s3_register("dbplyr::simulate_vars", "tbl_spark")
+    vctrs::s3_register("dbplyr::simulate_vars_is_typed", "tbl_spark")
+  }
+}
+
 utils::globalVariables(".")